refactor: combine payload creation into one class
