+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/injection/SessionModule.kt
@@ -1,12 +1,11 @@
 package io.embrace.android.embracesdk.injection
 
 import io.embrace.android.embracesdk.ndk.NativeModule
-import io.embrace.android.embracesdk.session.BackgroundActivityCollator
 import io.embrace.android.embracesdk.session.BackgroundActivityService
 import io.embrace.android.embracesdk.session.EmbraceBackgroundActivityService
 import io.embrace.android.embracesdk.session.EmbraceSessionService
+import io.embrace.android.embracesdk.session.PayloadMessageCollator
 import io.embrace.android.embracesdk.session.SessionHandler
-import io.embrace.android.embracesdk.session.SessionMessageCollator
 import io.embrace.android.embracesdk.session.SessionService
 import io.embrace.android.embracesdk.session.properties.EmbraceSessionProperties
 import io.embrace.android.embracesdk.session.properties.EmbraceSessionPropertiesService
@@ -18,8 +17,7 @@ internal interface SessionModule {
     val sessionHandler: SessionHandler
     val sessionService: SessionService
     val backgroundActivityService: BackgroundActivityService?
-    val sessionMessageCollator: SessionMessageCollator
-    val backgroundActivityCollator: BackgroundActivityCollator
+    val payloadMessageCollator: PayloadMessageCollator
     val sessionPropertiesService: SessionPropertiesService
 }
 
@@ -38,8 +36,8 @@ internal class SessionModuleImpl(
     workerThreadModule: WorkerThreadModule
 ) : SessionModule {
 
-    override val sessionMessageCollator: SessionMessageCollator by singleton {
-        SessionMessageCollator(
+    override val payloadMessageCollator: PayloadMessageCollator by singleton {
+        PayloadMessageCollator(
             essentialServiceModule.configService,
             essentialServiceModule.metadataService,
             dataContainerModule.eventService,
@@ -51,6 +49,8 @@ internal class SessionModuleImpl(
             nativeModule.nativeThreadSamplerService,
             dataCaptureServiceModule.breadcrumbService,
             essentialServiceModule.userService,
+            androidServicesModule.preferencesService,
+            initModule.spansService,
             initModule.clock
         )
     }
@@ -59,7 +59,6 @@ internal class SessionModuleImpl(
         SessionHandler(
             coreModule.logger,
             essentialServiceModule.configService,
-            androidServicesModule.preferencesService,
             essentialServiceModule.userService,
             essentialServiceModule.networkConnectivityService,
             essentialServiceModule.metadataService,
@@ -69,7 +68,7 @@ internal class SessionModuleImpl(
             sdkObservabilityModule.internalErrorService,
             essentialServiceModule.memoryCleanerService,
             deliveryModule.deliveryService,
-            sessionMessageCollator,
+            payloadMessageCollator,
             sessionProperties,
             initModule.clock,
             initModule.spansService,
@@ -107,26 +106,11 @@ internal class SessionModuleImpl(
                 essentialServiceModule.configService,
                 nativeModule.ndkService,
                 initModule.clock,
-                backgroundActivityCollator,
+                payloadMessageCollator,
                 workerThreadModule.backgroundWorker(WorkerName.PERIODIC_CACHE)
             )
         } else {
             null
         }
-    }
-
-    override val backgroundActivityCollator: BackgroundActivityCollator by singleton {
-        BackgroundActivityCollator(
-            essentialServiceModule.userService,
-            androidServicesModule.preferencesService,
-            dataContainerModule.eventService,
-            customerLogModule.logMessageService,
-            sdkObservabilityModule.internalErrorService,
-            dataCaptureServiceModule.breadcrumbService,
-            essentialServiceModule.metadataService,
-            dataContainerModule.performanceInfoService,
-            initModule.spansService,
-            initModule.clock
-        )
     }
 }
