+++ embrace-android-sdk/src/main/java/io/embrace/android/embracesdk/session/PayloadMessageCollator.kt
@@ -11,15 +11,18 @@ import io.embrace.android.embracesdk.config.ConfigService
 import io.embrace.android.embracesdk.event.EventService
 import io.embrace.android.embracesdk.event.LogMessageService
 import io.embrace.android.embracesdk.internal.clock.Clock
+import io.embrace.android.embracesdk.internal.spans.EmbraceAttributes
 import io.embrace.android.embracesdk.internal.spans.EmbraceSpanData
+import io.embrace.android.embracesdk.internal.spans.SpansService
+import io.embrace.android.embracesdk.internal.utils.Uuid
 import io.embrace.android.embracesdk.logging.InternalErrorService
 import io.embrace.android.embracesdk.payload.BetaFeatures
 import io.embrace.android.embracesdk.payload.Session
 import io.embrace.android.embracesdk.payload.SessionMessage
-import io.embrace.android.embracesdk.payload.UserInfo
+import io.embrace.android.embracesdk.prefs.PreferencesService
 import io.embrace.android.embracesdk.session.properties.EmbraceSessionProperties
 
-internal class SessionMessageCollator(
+internal class PayloadMessageCollator(
     private val configService: ConfigService,
     private val metadataService: MetadataService,
     private val eventService: EventService,
@@ -31,32 +34,47 @@ internal class SessionMessageCollator(
     private val nativeThreadSamplerService: NativeThreadSamplerService?,
     private val breadcrumbService: BreadcrumbService,
     private val userService: UserService,
+    private val preferencesService: PreferencesService,
+    private val spansService: SpansService,
     private val clock: Clock
 ) {
+
+    internal enum class PayloadType {
+        SESSION,
+        BACKGROUND_ACTIVITY
+    }
 
     /**
      * Builds a new session object. This should not be sent to the backend but is used
      * to populate essential session information (such as ID), etc
      */
     internal fun buildInitialSession(
-        id: String,
+        payloadType: PayloadType,
         coldStart: Boolean,
         startType: Session.LifeEventType,
         startTime: Long,
-        sessionNumber: Int,
-        userInfo: UserInfo?,
-        sessionProperties: Map<String, String>
-    ): Session = Session(
-        sessionId = id,
-        startTime = startTime,
-        number = sessionNumber,
-        appState = Session.APPLICATION_STATE_FOREGROUND,
-        isColdStart = coldStart,
-        startType = startType,
-        properties = sessionProperties,
-        messageType = Session.MESSAGE_TYPE_END,
-        user = userInfo
-    )
+        sessionProperties: Map<String, String>?
+    ): Session {
+        val appState = when (payloadType) {
+            PayloadType.SESSION -> Session.APPLICATION_STATE_FOREGROUND
+            PayloadType.BACKGROUND_ACTIVITY -> Session.APPLICATION_STATE_BACKGROUND
+        }
+        val sessionNumber = when (payloadType) {
+            PayloadType.SESSION -> preferencesService.incrementAndGetSessionNumber()
+            PayloadType.BACKGROUND_ACTIVITY -> preferencesService.incrementAndGetBackgroundActivityNumber()
+        }
+        return Session(
+            sessionId = Uuid.getEmbUuid(),
+            startTime = startTime,
+            isColdStart = coldStart,
+            messageType = Session.MESSAGE_TYPE_END,
+            appState = appState,
+            startType = startType,
+            user = captureDataSafely(userService::loadUserInfoFromDisk),
+            number = sessionNumber,
+            properties = sessionProperties,
+        )
+    }
 
     /**
      * Builds a fully populated session message. This can be sent to the backend (or stored
@@ -171,5 +189,98 @@ internal class SessionMessageCollator(
             breadcrumbs = breadcrumbs,
             spans = spans
         )
+    }
+
+    /**
+     * Creates a background activity stop message.
+     */
+    fun createBackgroundActivityEndMessage(
+        activity: Session,
+        endTime: Long,
+        endType: Session.LifeEventType?,
+        crashId: String?
+    ): Session {
+        val startTime = activity.startTime ?: 0
+        return activity.copy(
+            appState = Session.APPLICATION_STATE_BACKGROUND,
+            messageType = Session.MESSAGE_TYPE_END,
+            endTime = endTime,
+            eventIds = captureDataSafely {
+                eventService.findEventIdsForSession(
+                    startTime,
+                    endTime
+                )
+            },
+            infoLogIds = captureDataSafely { logMessageService.findInfoLogIds(startTime, endTime) },
+            warningLogIds = captureDataSafely {
+                logMessageService.findWarningLogIds(
+                    startTime,
+                    endTime
+                )
+            },
+            errorLogIds = captureDataSafely { logMessageService.findErrorLogIds(startTime, endTime) },
+            infoLogsAttemptedToSend = captureDataSafely(logMessageService::getInfoLogsAttemptedToSend),
+            warnLogsAttemptedToSend = captureDataSafely(logMessageService::getWarnLogsAttemptedToSend),
+            errorLogsAttemptedToSend = captureDataSafely(logMessageService::getErrorLogsAttemptedToSend),
+            exceptionError = captureDataSafely(internalErrorService::currentExceptionError),
+            lastHeartbeatTime = endTime,
+            endType = endType,
+            unhandledExceptions = captureDataSafely(logMessageService::getUnhandledExceptionsSent),
+            crashReportId = crashId
+        )
+    }
+
+    /**
+     * Create the background session message with the current state of the background activity.
+     *
+     * @param backgroundActivity      the current state of a background activity
+     * @param isBackgroundActivityEnd true if the message is being built for the termination of the background activity
+     * @return a background activity message for backend
+     */
+    fun buildBgActivityMessage(
+        backgroundActivity: Session?,
+        isBackgroundActivityEnd: Boolean
+    ): SessionMessage? {
+        if (backgroundActivity != null) {
+            val startTime = backgroundActivity.startTime ?: 0L
+            val endTime = backgroundActivity.endTime ?: clock.now()
+            val isCrash = backgroundActivity.crashReportId != null
+            val breadcrumbs = captureDataSafely {
+                when {
+                    isBackgroundActivityEnd -> breadcrumbService.flushBreadcrumbs()
+                    else -> breadcrumbService.getBreadcrumbs(startTime, endTime)
+                }
+            }
+            val spans: List<EmbraceSpanData>? = captureDataSafely {
+                when {
+                    isBackgroundActivityEnd -> {
+                        val appTerminationCause = when {
+                            isCrash -> EmbraceAttributes.AppTerminationCause.CRASH
+                            else -> null
+                        }
+                        spansService.flushSpans(appTerminationCause)
+                    }
+                    else -> spansService.completedSpans()
+                }
+            }
+
+            return SessionMessage(
+                session = backgroundActivity,
+                userInfo = backgroundActivity.user,
+                appInfo = captureDataSafely(metadataService::getAppInfo),
+                deviceInfo = captureDataSafely(metadataService::getDeviceInfo),
+                performanceInfo = captureDataSafely {
+                    performanceInfoService.getSessionPerformanceInfo(
+                        startTime,
+                        endTime,
+                        java.lang.Boolean.TRUE == backgroundActivity.isColdStart,
+                        null
+                    )
+                },
+                breadcrumbs = breadcrumbs,
+                spans = spans
+            )
+        }
+        return null
     }
 }
