+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/fakes/injection/FakeSessionModule.kt
@@ -3,10 +3,9 @@ package io.embrace.android.embracesdk.fakes.injection
 import io.embrace.android.embracesdk.FakeSessionPropertiesService
 import io.embrace.android.embracesdk.FakeSessionService
 import io.embrace.android.embracesdk.injection.SessionModule
-import io.embrace.android.embracesdk.session.BackgroundActivityCollator
 import io.embrace.android.embracesdk.session.BackgroundActivityService
+import io.embrace.android.embracesdk.session.PayloadMessageCollator
 import io.embrace.android.embracesdk.session.SessionHandler
-import io.embrace.android.embracesdk.session.SessionMessageCollator
 import io.embrace.android.embracesdk.session.SessionService
 import io.embrace.android.embracesdk.session.properties.SessionPropertiesService
 
@@ -18,9 +17,6 @@ internal class FakeSessionModule(
     override val sessionHandler: SessionHandler
         get() = TODO("Not yet implemented")
 
-    override val sessionMessageCollator: SessionMessageCollator
-        get() = TODO("Not yet implemented")
-
-    override val backgroundActivityCollator: BackgroundActivityCollator
+    override val payloadMessageCollator: PayloadMessageCollator
         get() = TODO("Not yet implemented")
 }
