+++ embrace-android-sdk/src/test/java/io/embrace/android/embracesdk/session/SessionModuleImplTest.kt
@@ -46,7 +46,7 @@ internal class SessionModuleImplTest {
         )
         assertNotNull(module.sessionHandler)
         assertNotNull(module.sessionService)
-        assertNotNull(module.sessionMessageCollator)
+        assertNotNull(module.payloadMessageCollator)
         assertNotNull(module.sessionPropertiesService)
         assertNull(module.backgroundActivityService)
     }
@@ -69,7 +69,7 @@ internal class SessionModuleImplTest {
         )
         assertNotNull(module.sessionHandler)
         assertNotNull(module.sessionService)
-        assertNotNull(module.sessionMessageCollator)
+        assertNotNull(module.payloadMessageCollator)
         assertNotNull(module.sessionPropertiesService)
         assertNotNull(module.backgroundActivityService)
     }
